Receiver reads settings from file. Changed modules
