+++ webinterface/templates/configuration.html
@@ -15,6 +15,7 @@
             <table class="table is-narrow is-hoverable is-fullwidth">
                 <col width="200px">
                 <tr><td>Appliance Name:</td><td>{{config["appliance_name"]}}</td></tr>
+                <tr><td>Port:</td><td>{{config["port"]}}</td></tr>                
                 <tr><td>Version:</td><td>{{hermes_version}}</td></tr>
                 <tr><td>Operating System:</td><td>{{os_string}}</td></tr>
             </table>
